+++ src/app/api/config.service.ts
@@ -3,7 +3,7 @@ import { Injectable } from '@angular/core';
 import { Observable } from 'rxjs/Observable';
 
 import { environment } from 'environments/environment';
-import { AdsharesAddress, DepositInfo } from 'models/settings.model';
+import { AdsharesAddress, DepositInfo, NowPaymentsInit } from 'models/settings.model';
 
 @Injectable()
 export class ApiConfigService {
@@ -17,5 +17,9 @@ export class ApiConfigService {
 
   depositInfo(): Observable<DepositInfo> {
     return this.http.get<DepositInfo>(`${environment.apiUrl}/deposit-info`);
+  }
+
+  nowPaymentsInit(amount: number): Observable<NowPaymentsInit> {
+    return this.http.get<NowPaymentsInit>(`${environment.apiUrl}/now-payments/init?amount=`+amount);
   }
 }