+++ src/app/models/user.model.ts
@@ -11,6 +11,12 @@ interface ExchangeRate {
   value: number;
 }
 
+interface NowPaymentsInfo {
+  minAmount: number;
+  exchangeRate: number;
+  currency: string;
+}
+
 interface User {
   id: number
   email: string;
@@ -26,6 +32,7 @@ interface User {
   exchangeRate: ExchangeRate | null;
   uuid: string;
   referralId?: string;
+  nowPayments: NowPaymentsInfo | null;
 }
 
 interface LocalStorageUser extends User {
@@ -40,4 +47,4 @@ interface UserRoles extends User {
   advertiser: string;
 }
 
-export { UserAdserverWallet, User, LocalStorageUser, ExchangeRate };
+export { UserAdserverWallet, User, LocalStorageUser, ExchangeRate, NowPaymentsInfo };