+++ src/app/settings/billing/user-wallet/user-wallet.component.ts
@@ -7,8 +7,9 @@ import { appSettings } from 'app-settings';
 import { SessionService } from "app/session.service";
 import {AppState} from "models/app-state.model";
 import {Store} from "@ngrx/store";
-import { UserAdserverWallet } from "models/user.model";
+import { NowPaymentsInfo, UserAdserverWallet } from "models/user.model";
 import { CODE, CRYPTO } from "common/utilities/consts";
+import { NowPaymentsDialogComponent } from "common/dialog/now-payments-dialog/now-payments-dialog.component";
 
 @Component({
   selector: 'app-user-wallet',
@@ -18,6 +19,7 @@ import { CODE, CRYPTO } from "common/utilities/consts";
 export class UserWalletComponent extends HandleSubscription implements OnInit {
   faqLink = appSettings.FAQ_LINK;
   wallet: UserAdserverWallet;
+  nowPayments: NowPaymentsInfo;
   crypto: string = CRYPTO;
   code: string = CODE;
 
@@ -37,10 +39,18 @@ export class UserWalletComponent extends HandleSubscription implements OnInit {
     this.dialog.open(WithdrawFundsDialogComponent);
   }
 
+  openNowPaymentsDialog() {
+    this.dialog.open(NowPaymentsDialogComponent);
+  }
+
   ngOnInit() {
     this.store.select('state', 'user', 'data', 'adserverWallet')
       .subscribe((wallet: UserAdserverWallet) => {
         this.wallet = wallet;
+      });
+    this.store.select('state', 'user', 'data', 'nowPayments')
+      .subscribe((info: NowPaymentsInfo) => {
+        this.nowPayments = info;
       });
   }
 }